Change persistenceContainer() and saveContext() to application.swift
